# Change the representation of IP to a pair of uint64s.