@josharian 2020-12-24
Fuzz testing found a bug. See https://github.com/inetaf/netaddr/pull/66.

Time for me to get the fuzz function committed so that you can use it too. PR coming shortly.