@danderson 2020-12-24
Fixed your feedback and the failure found by fuzz testing (the test for Is4in6 was incorrect).