@josharian 2020-12-24
I'm wrapping up for the day and week. Looking forward to reviewing next week. This is lovely.