@danderson 2020-12-25
For reference, https://github.com/golang/go/issues/41663 is the reason that I didn't get the same performance improvement when I kept [16]byte and did a BigEndian read-twiddle-write where needed. You might expect the resulting code to be the same, but it breaks down into byte-wise twiddling.

Whereas casting the bytes into uint64 from the get-go compiles efficiently in both sites: the BigEndian read/write are a MOVQ+BSWAPQ, and the twiddling ops are 64-bit register math.