@danderson 2020-12-25
Managed to dump SSA, and as expected the compiler inlines, but doesn't do any intelligent coalescing. The check `v4(0) == 169 && v4(1) == 254` ends up doing:

```
MOVQ "".ip+8(SP), AX  # Load all of ip.lo into A
MOVQ AX, CX # Copy A to C
SHRQ $24, AX # Shift A right 24
CMPB AX, $-87 # v4(0) == 169
JNE 19
SHRQ $16, CX # Shift C right 16
CMPB CX, $-2 # v4(1) == 254
```

Compare to `uint16(ip.lo>>16) == 0xa9fe`:

```
MOVQ "".ip+8(SP), AX # Load all of ip.lo into A
SHRQ $16, AX # Shift A right 16
CMPW AX, $-22018 # Compare lower 16 bits of register to 0xa9fe
```

So, the generated code is definitely worse, and in line with what I saw when I tried to use encoding/binary to Uint64+twiddle+PutUint64. Based on that, I would prefer to continue operating on larger chunks of the address, rather than break it down bytewise. Maybe something like `v4ab() uint16`, and similar custom accessors for each different slice of IP that we need to examine? That would still make the high-level functions more readable, at the expense of a bit more duplication of bit twiddlers.