@bradfitz 2020-12-25
> Based on that, I would prefer to continue operating on larger chunks of the address, rather than break it down bytewise. 

I'd prefer to err on the side of explicit & readable code until there's a performance reason to micro-optimize `IsLinkLocalUnicast`.

Out of curiosity, does Go do any better at `ip.v4(0)<<8|ip.v4(1) == (169<<8)|254` ?