@danderson 2020-12-25
Not by much. The compare becomes at 16-byte CMPW, but the load is still dumb (loads ip.lo twice into two different registers, shifts each one by a different amount, then ORs them together).

Keeping at bytewise loads for now. Will see in a followup if I can make it generate slightly less rubbish code without compromising readability.